+++ src/components/Feed.js
@@ -17,11 +17,14 @@ const Feed = () => {
   }
 
   const deleteChirp = async (chirpId) => {
+  if (window.confirm("Are you sure you want to delete this chirp?")) {
     await ApiRequest.deleteChirp(
       { 'chirp_id': chirpId}
     );
     getChirps();
-  };
+  }
+};
+
 
   useEffect(() => {
     getChirps()
